Pages 15 to 21
Justin had done pages 8 to 14, I have added pages 15 to 21. Grant, you
will need to do pages 22 to 28. This gives each of us 7 pages of data to
enter. Some notes to be aware of: Tryon county no longer exsist so it
needs to be entered in as Montogmery. This is true for Charlotte County
as will. It is now Washington County. Dutchys county is actually
Dutchees county. There is also a an unidntified county called ?Iuiens?.
The town connected to it is Newtown, and that town is in prsent day
Chemung county. We have to have the current county names inputed in
order for this to work on th current GIS New York data.